% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{annot.chr21}
\alias{annot.chr21}
\title{annot.chr21: Annotation information for chromosome 21, hg38 genome}
\format{
a \code{GRangesList} object with two elements returned
by \code{\link{getAnnot}}. The first
contains CpG category information in the first element (optional)
coding gene sequence information in the second element (optional).
At least one of these elements needs to be non-null in order for 
any annotation to be plotted, but it is not necessary to contain
both.
}
\source{
Obtained from running
\code{annoTrack} function and then subsetting the results to 
only include chromosome 21. A script which executes these steps 
and constructs the \code{annot.chr21} 
 object may be found in \file{inst/scripts/get_annot.chr21.R}
}
\usage{
data(annot.chr21)
}
\description{
This is the annotation information returned from 
\code{\link{getAnnot}}, subsetted for chromosome 21 for convenience
in running the examples. The annotation is obtained using the 
\code{annotatr} package.
}
\examples{
data(annot.chr21)
}
\keyword{datasets}
