% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnot.R
\name{build_cpg_annots2}
\alias{build_cpg_annots2}
\title{A helper function to build CpG related annotations.}
\usage{
build_cpg_annots2(genome = builtin_genomes2(),
  annotations = builtin_annotations2())
}
\arguments{
\item{genome}{The genome assembly.}

\item{annotations}{A character vector with entries of the form 
\code{[genome]_cpg_{islands,shores,shelves,inter}}.}
}
\value{
A list of \code{GRanges} objects.
}
\description{
Using the \code{AnnotationHub} package, extract CpG island track 
for the appropriate \code{genome} and construct the shores, shelves, and 
interCGI annotations as desired. ((This function is a modified version
of that found in the annotatr package, simply to include the hg18 version
of the genome for CpG categories))
}
