% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnot.R
\name{builtin_annotations2}
\alias{builtin_annotations2}
\title{Function listing which annotations are available.
Modified from annotatr to include hg18}
\usage{
builtin_annotations2()
}
\value{
A character vector of available annotations.
}
\description{
This includes the shortcuts. The \code{expand_annotations()} function helps
handle the shortcuts.
}
\examples{
builtin_annotations2()

}
