% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnot.R
\name{builtin_genomes2}
\alias{builtin_genomes2}
\title{Function returning supported TxDb.* genomes
modified from annotatr to include hg18}
\usage{
builtin_genomes2()
}
\value{
A character vector of genomes for supported TxDb.* packages
}
\description{
Function returning supported TxDb.* genomes
modified from annotatr to include hg18
}
\examples{
builtin_genomes2()

}
