% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnot.R
\name{check_annotations2}
\alias{check_annotations2}
\title{Function to check for valid annotations
Modified from annotatr to include hg18}
\usage{
check_annotations2(annotations)
}
\arguments{
\item{annotations}{A character vector of annotations 
possibly using the shortcuts}
}
\value{
If all the checks on the annotations pass, 
returns NULL to allow code to move forward.
}
\description{
Gives errors if any annotations are not in builtin_annotations2() 
(and they are not in the required custom format), basicgenes are used,
 or the genome prefixes are not the same for all annotations.
}
