% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_plotting_functions.R
\name{dmrPlotAnnotations}
\alias{dmrPlotAnnotations}
\title{Add annotations to DMR plots}
\usage{
dmrPlotAnnotations(gr, annoTrack)
}
\arguments{
\item{gr}{a \code{GRanges} object that contains the DMRs to be plotted}

\item{annoTrack}{a \code{SimpleGenomicRangesList} object with two elements. 
The first contains CpG category information in the first element (optional)
coding gene sequence information in the second element (optional).
At least one of these elements needs to be non-null in order for 
any annotation to be plotted, but it is not necessary to contain
both.}
}
\value{
None
}
\description{
Function to add visual representation of CpG categories 
and/or coding 
sequences to DMR plot
}
\details{
An internal function that takes an annotation \code{GRangesList} 
object that
contains CpG category information in the first element (optional) and / or
coding gene sequence information in the second element (optional). If neither
of these are present, then nothing will be plotted.
}
