% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLoci.R
\name{filterLoci}
\alias{filterLoci}
\title{Filter loci for coverage}
\usage{
filterLoci(bs, minCoverage = 1)
}
\arguments{
\item{bs}{a \code{BSseq} object}

\item{minCoverage}{the minimum coverage value allowed for any given loci.
The default value (recommended) is 1. This means that any loci with zero
reads mapping to it in at least one sample are removed.}
}
\value{
a \code{BSseq} object with the offending loci removed.
}
\description{
Helper function to remove loci that have low coverage.
}
\examples{

# load example data 
data(BS.chr21)

BS.chr21 <- filterLoci(BS.chr21)

}
