% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnot.R
\name{get_orgdb_name2}
\alias{get_orgdb_name2}
\title{Function to get correct org.* package name based on genome
modified from annotatr to include hg18}
\usage{
get_orgdb_name2(genome = builtin_genomes2())
}
\arguments{
\item{genome}{A string giving the genome assembly.}
}
\value{
A string giving the correct org for org.db packages. e.g. hg19 -> Hs.
}
\description{
Function to get correct org.* package name based on genome
modified from annotatr to include hg18
}
