% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnot.R
\name{get_txdb_name2}
\alias{get_txdb_name2}
\title{Function to get correct TxDb.* package name based on genome
modified from annotatr to include hg18}
\usage{
get_txdb_name2(genome = builtin_genomes2())
}
\arguments{
\item{genome}{A string giving the genome assembly.}
}
\value{
A string giving the name of the correct TxDb.* package name based 
on \code{genome}.
}
\description{
Function to get correct TxDb.* package name based on genome
modified from annotatr to include hg18
}
