% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnot.R
\name{getAnnot}
\alias{getAnnot}
\title{Retrieve annotation information}
\usage{
getAnnot(genomeName)
}
\arguments{
\item{genomeName}{a character object that indicates which organism is 
under study. Use the function \code{annotatr:::builtin_genomes()} to see
a character vector of available genome names to choose from (see 
\code{annotatr} documentation for more details).}
}
\description{
Uses the \code{annotatr} package to retrieve annotation information (
CpG category and gene coding sequences) for the \code{annoTrack} argument
of \code{\link{plotDMRs}}. Allows for 5 
re-tries if download fails (to allow for a spotty internet connection).
}
