% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmpiricalDistribution.R
\name{plotEmpiricalDistribution}
\alias{plotEmpiricalDistribution}
\title{Plot the empirical distribution of the methylation beta vals}
\usage{
plotEmpiricalDistribution(bs, plotFile, tiss1, tiss2, time = NULL,
  time2 = NULL, matchCovariate = NULL, genomeName = NULL)
}
\arguments{
\item{bs}{a BSseq object}

\item{plotFile}{a character string specifying the filename (and full path,
or else file will be saved to the current working directory) for the 
pdf file that will contain the output plot.}

\item{tiss1}{character indicating the label of population 1}

\item{tiss2}{character indicating the lable of population 2}

\item{time}{character indicating a secondary label of population 1 (e.g.
time point if each population is measured at multiple time points)}

\item{time2}{character indicating a secondary label of population 2 (e.g.
time point if each population is measured at multiple time points)}

\item{matchCovariate}{character specifying the column name of the
\code{pData} slot of the BSseq object to include in the plot legend.}

\item{genomeName}{character specifying the genome name if plots of
beta values by CpG category should be made. See package \code{annotatr}
for details of available genomes.}
}
\value{
None. Produces a plot saved to a pdf file with filename 
\code{plotFile}
}
\description{
Uses ggplot2 to plot smoothed density histograms of methylation
proportions (beta values)
}
