% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmrseq.R
\name{dmrseq}
\alias{dmrseq}
\title{Main function for detecting and evaluating significance of DMRs.}
\usage{
dmrseq(bs, testCovariate, adjustCovariate = NULL, cutoff = 0.1,
  minNumRegion = 5, smooth = TRUE, bpSpan = 1000, minInSpan = 30,
  maxGapSmooth = 2500, maxGap = 1000, verbose = TRUE, workers = NULL,
  sampleIndex = seq(1, nrow(pData(bs))), maxPerms = 20,
  matchCovariate = NULL)
}
\arguments{
\item{bs}{bsseq object containing the methylation values as well as the
phenotype matrix that contains sample level covariates}

\item{testCovariate}{integer value or vector indicating which of columns of
\code{pData(bs)} contains the covariate of interest that will be tested 
for association of methylation levels. This is used to construct the 
design matrix for the test statistic calculation.}

\item{adjustCovariate}{an (optional) integer value or vector indicating
which of the columns of \code{pData(bs)} will be adjusted for when 
testing for the association of methylation value with the 
\code{testCovariate}. If not NULL (default), then this is also used to 
construct the design matrix for the test statistic calculation.}

\item{cutoff}{scalar value that represents the absolute value (or a vector 
of two numbers representing a lower and upper bound) for the cutoff of 
the single CpG coefficient that is used to discover 
candidate regions. Default value is 0.10.}

\item{minNumRegion}{positive integer that represents the minimum number of
nucleotides to consider for a candidate region. Default value is 5.}

\item{smooth}{logical value that indicates whether or not to smooth the 
CpG level signal when discovering candidate regions.
Defaults to TRUE.}

\item{bpSpan}{a positive integer that represents the length in basepairs
of the smoothing span window if \code{smooth} is TRUE.  Default value is 
1000.}

\item{minInSpan}{positive integer that represents the minimum number of
CpGs in a smoothing span window if \code{smooth} is TRUE.  
Default value is 30.}

\item{maxGapSmooth}{integer value representing maximum number of basepairs
in between neighboring CpGs to be included in the same 
cluster when performing smoothing (should generally be larger than
\code{maxGap})}

\item{maxGap}{integer value representing maximum number of basepairs in 
between neighboring CpGs to be included in the same DMR.}

\item{verbose}{logical value that indicates whether progress messages
should be printed to stdout. Defaults value is TRUE.}

\item{workers}{positive integer that represents the number of cores to 
use if parallelization is desired of the smoothing step.}

\item{sampleIndex}{vector of positive integers that represents the 
index of samples to use in the analysis (corresponds to sample ordering
in \code{bs} (e.g. the order of the rows in \code{pData(bs)}. 
Default is to use all samples in the object.}

\item{maxPerms}{a positive integer that represents the maximum number 
of permutations that will be used to generate the global null 
distribution of test statistics.  Default value is 20.}

\item{matchCovariate}{an (optional) integer value or vector indicating
which of the columns of \code{pData(bs)} will be blocked for when 
constructing the permutations in order to
test for the association of methylation value with the 
\code{testCovariate}. Blocking means that only permutations with balanced
composition of \code{testCovariate} values will be used (for example if
you have samples from different gender and this is not your covariate of
interest, 
it is recommended to use gender as a matching covariate to avoid one 
of the permutations testing entirely males versus females; this violates
the null hypothesis and will decrease power).
If not NULL (default), then no blocking is performed.}
}
\value{
a data.frame that contains the results of the inference. The
   data.frame contains one row for each candidate region, and 
   10 columns, in the following order: 1. chr = 
   chromosome, 2. start = 
   start basepair position of the region, 3. end = end basepair position
   of the region,
   4. indexStart = the index of the region's first CpG, 
   5. indexEnd = the index of the region's last CpG,
   6. L = the number of CpGs contained in the region,
   7. beta = the coefficient value for the condition difference,
   8. stat = the test statistic for the condition difference,
   9. pval = the permutation p-value for the significance of the test
   statistic, and 10. qval = the q-value for the test statistic (adjustment
   for multiple comparisons to control false discovery rate).
}
\description{
Performs a three-step approach that (1) detects candidate regions, 
(2) scores candidate regions with an exchangeable (across the genome)
statistic, and (3) evaluates statistical significance using a 
permuation test on the pooled null distribution of scores.
}
\keyword{inference}
